Implemented TRuleSendMessageAction class
