+++ IceWarpLib/IceWarpLib.Objects/IceWarpLib.Objects.csproj
@@ -156,23 +156,24 @@
     <Compile Include="Rpc\Classes\Property\TPropertyVal.cs" />
     <Compile Include="Rpc\Classes\Property\TPropertyValue.cs" />
     <Compile Include="Rpc\Classes\Property\TPropertyValueList.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleAction.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleAllCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleConditions.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleDNSBLCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleHasAttachmentCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Actions\TRuleAction.cs" />
+    <Compile Include="Rpc\Classes\Rule\Actions\TRuleSendMessageAction.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleAllCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleConditions.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleDNSBLCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleHasAttachmentCondition.cs" />
     <Compile Include="Rpc\Classes\Rule\TRuleInfo.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleIsSizeCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleIsSpamCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleLocalTimeCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRulePriorityCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleSenderRecipientCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleIsSizeCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleIsSpamCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleLocalTimeCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRulePriorityCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleSenderRecipientCondition.cs" />
     <Compile Include="Rpc\Classes\Rule\TRulesListFilter.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleSMTPAuthCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleSomeWordsCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleSpamScoreCondition.cs" />
-    <Compile Include="Rpc\Classes\Rule\TRuleTrustedSessionCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleSMTPAuthCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleSomeWordsCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleSpamScoreCondition.cs" />
+    <Compile Include="Rpc\Classes\Rule\Conditions\TRuleTrustedSessionCondition.cs" />
     <Compile Include="Rpc\Enums\AccountType.cs" />
     <Compile Include="Rpc\Enums\TAccountPropertySet.cs" />
     <Compile Include="Rpc\Enums\TActivationKeyType.cs" />
