+++ IceWarpLib/IceWarpLib.Objects/Rpc/Classes/Rule/Conditions/TRuleIsSpamCondition.cs
@@ -1,14 +1,14 @@
 ﻿using System.Xml;
 using IceWarpLib.Objects.Helpers;
 
-namespace IceWarpLib.Objects.Rpc.Classes.Rule
+namespace IceWarpLib.Objects.Rpc.Classes.Rule.Conditions
 {
     /// <summary>
-    /// Has attachment condition type
+    /// Is spam condition type
     /// </summary>
-    public class TRuleHasAttachmentCondition : TRuleCondition
+    public class TRuleIsSpamCondition : TRuleCondition
     {
-        public TRuleHasAttachmentCondition()
+        public TRuleIsSpamCondition()
         {
         }
 
@@ -16,7 +16,7 @@ namespace IceWarpLib.Objects.Rpc.Classes.Rule
         /// Creates new instance from an XML node. See <see cref="XmlNode"/> for more information.
         /// </summary>
         /// <param name="node">The Xml node. See <see cref="XmlNode"/> for more information.</param>
-        public TRuleHasAttachmentCondition(XmlNode node)
+        public TRuleIsSpamCondition(XmlNode node)
         {
             if (node != null)
             {
