+++ IceWarpLib/IceWarpLib.Objects/Rpc/Classes/Rule/Conditions/TRuleSMTPAuthCondition.cs
@@ -1,20 +1,14 @@
 ﻿using System.Xml;
 using IceWarpLib.Objects.Helpers;
-using IceWarpLib.Objects.Rpc.Enums;
 
-namespace IceWarpLib.Objects.Rpc.Classes.Rule
+namespace IceWarpLib.Objects.Rpc.Classes.Rule.Conditions
 {
     /// <summary>
-    /// Priority condition type
+    /// Bayes score condition type Antivirus condition type SMTP Auth condition type
     /// </summary>
-    public class TRulePriorityCondition : TRuleCondition
+    public class TRuleSMTPAuthCondition : TRuleCondition
     {
-        /// <summary>
-        /// Represents class property TRulePriorityCondition.Priority
-        /// </summary>
-        public TRulePriorityType Priority { get; set; }
-
-        public TRulePriorityCondition()
+        public TRuleSMTPAuthCondition()
         {
         }
 
@@ -22,12 +16,11 @@ namespace IceWarpLib.Objects.Rpc.Classes.Rule
         /// Creates new instance from an XML node. See <see cref="XmlNode"/> for more information.
         /// </summary>
         /// <param name="node">The Xml node. See <see cref="XmlNode"/> for more information.</param>
-        public TRulePriorityCondition(XmlNode node)
+        public TRuleSMTPAuthCondition(XmlNode node)
         {
             if (node != null)
             {
                 ProcessNode(node);
-                Priority = (TRulePriorityType)Extensions.GetNodeInnerTextAsInt(node.GetSingleNode("Priority"));
             }
         }
 
@@ -36,7 +29,6 @@ namespace IceWarpLib.Objects.Rpc.Classes.Rule
             XmlElement element = XmlHelper.CreateElement(doc, name);
 
             AppendBaseElements(element);
-            XmlHelper.AppendTextElement(element, "Priority", Priority);
 
             return element;
         }
