--- IceWarpLib/IceWarpLib.Objects/Rpc/Classes/Rule/TRuleAllCondition.cs
@@ -1,33 +1,0 @@
-﻿using System.Xml;
-using IceWarpLib.Objects.Helpers;
-
-namespace IceWarpLib.Objects.Rpc.Classes.Rule
-{
-    public class TRuleAllCondition : TRuleCondition
-    {
-        public TRuleAllCondition()
-        {
-        }
-
-        /// <summary>
-        /// Creates new instance from an XML node. See <see cref="XmlNode"/> for more information.
-        /// </summary>
-        /// <param name="node">The Xml node. See <see cref="XmlNode"/> for more information.</param>
-        public TRuleAllCondition(XmlNode node)
-        {
-            if (node != null)
-            {
-                ProcessNode(node);
-            }
-        }
-
-        public override XmlElement BuildXmlElement(XmlDocument doc, string name)
-        {
-            XmlElement element = XmlHelper.CreateElement(doc, name);
-
-            AppendBaseElements(element);
-
-            return element;
-        }
-    }
-}
