+++ IceWarpLib/IceWarpLib.UnitTests/IceWarpObjects/Rpc/Classes/Rule/Conditions/TRuleAllCondition_Test.cs
@@ -1,14 +1,14 @@
-﻿using IceWarpLib.Objects.Rpc.Classes.Rule;
+﻿using IceWarpLib.Objects.Rpc.Classes.Rule.Conditions;
 using IceWarpLib.Objects.Rpc.Enums;
 using NUnit.Framework;
 
-namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
+namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule.Conditions
 {
-    public class TRuleHasAttachmentCondition_Test : BaseTest
+    public class TRuleAllCondition_Test : BaseTest
     {
         private string _xml = @"
 <custom xmlns=""admin:iq:rpc"">
-    <classname>trulehasattachmentcondition</classname>
+    <classname>truleallcondition</classname>
     <conditiontype>0</conditiontype>
     <operatorand>0</operatorand>
     <logicalnot>0</logicalnot>
@@ -17,18 +17,18 @@ namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
 </custom>".TrimStart();
 
         [Test]
-        public void TRuleHasAttachmentCondition()
+        public void TRuleAllCondition()
         {
-            var testClass = new TRuleHasAttachmentCondition();
+            var testClass = new TRuleAllCondition();
 
             var testXml = ToFormattedXml(testClass);
             Assert.AreEqual(_xml, testXml);
         }
 
         [Test]
-        public void TRuleHasAttachmentCondition_BuildXmlElement()
+        public void TRuleAllCondition_BuildXmlElement()
         {
-            var testClass = new TRuleHasAttachmentCondition(GetXmlNode(_xml));
+            var testClass = new TRuleAllCondition(GetXmlNode(_xml));
 
             Assert.AreEqual(TRuleConditionType.None, testClass.ConditionType);
         }
