+++ IceWarpLib/IceWarpLib.UnitTests/IceWarpObjects/Rpc/Classes/Rule/Conditions/TRuleConditions_Test.cs
@@ -1,9 +1,9 @@
 ﻿using System.Linq;
-using IceWarpLib.Objects.Rpc.Classes.Rule;
+using IceWarpLib.Objects.Rpc.Classes.Rule.Conditions;
 using IceWarpLib.Objects.Rpc.Enums;
 using NUnit.Framework;
 
-namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
+namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule.Conditions
 {
     public class TRuleConditions_Test : BaseTest
     {
