+++ IceWarpLib/IceWarpLib.UnitTests/IceWarpObjects/Rpc/Classes/Rule/Conditions/TRuleIsSpamCondition_Test.cs
@@ -1,14 +1,14 @@
-﻿using IceWarpLib.Objects.Rpc.Classes.Rule;
+﻿using IceWarpLib.Objects.Rpc.Classes.Rule.Conditions;
 using IceWarpLib.Objects.Rpc.Enums;
 using NUnit.Framework;
 
-namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
+namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule.Conditions
 {
-    public class TRuleSMTPAuthCondition_Test : BaseTest
+    public class TRuleIsSpamCondition_Test : BaseTest
     {
         private string _xml = @"
 <custom xmlns=""admin:iq:rpc"">
-    <classname>trulesmtpauthcondition</classname>
+    <classname>truleisspamcondition</classname>
     <conditiontype>0</conditiontype>
     <operatorand>0</operatorand>
     <logicalnot>0</logicalnot>
@@ -17,18 +17,18 @@ namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
 </custom>".TrimStart();
 
         [Test]
-        public void TRuleSMTPAuthCondition()
+        public void TRuleIsSpamCondition()
         {
-            var testClass = new TRuleSMTPAuthCondition();
+            var testClass = new TRuleIsSpamCondition();
 
             var testXml = ToFormattedXml(testClass);
             Assert.AreEqual(_xml, testXml);
         }
 
         [Test]
-        public void TRuleSMTPAuthCondition_BuildXmlElement()
+        public void TRuleIsSpamCondition_BuildXmlElement()
         {
-            var testClass = new TRuleSMTPAuthCondition(GetXmlNode(_xml));
+            var testClass = new TRuleIsSpamCondition(GetXmlNode(_xml));
 
             Assert.AreEqual(TRuleConditionType.None, testClass.ConditionType);
         }
