+++ IceWarpLib/IceWarpLib.UnitTests/IceWarpObjects/Rpc/Classes/Rule/Conditions/TRuleSMTPAuthCondition_Test.cs
@@ -1,14 +1,14 @@
-﻿using IceWarpLib.Objects.Rpc.Classes.Rule;
+﻿using IceWarpLib.Objects.Rpc.Classes.Rule.Conditions;
 using IceWarpLib.Objects.Rpc.Enums;
 using NUnit.Framework;
 
-namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
+namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule.Conditions
 {
-    public class TRuleAllCondition_Test : BaseTest
+    public class TRuleSMTPAuthCondition_Test : BaseTest
     {
         private string _xml = @"
 <custom xmlns=""admin:iq:rpc"">
-    <classname>truleallcondition</classname>
+    <classname>trulesmtpauthcondition</classname>
     <conditiontype>0</conditiontype>
     <operatorand>0</operatorand>
     <logicalnot>0</logicalnot>
@@ -17,18 +17,18 @@ namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
 </custom>".TrimStart();
 
         [Test]
-        public void TRuleAllCondition()
+        public void TRuleSMTPAuthCondition()
         {
-            var testClass = new TRuleAllCondition();
+            var testClass = new TRuleSMTPAuthCondition();
 
             var testXml = ToFormattedXml(testClass);
             Assert.AreEqual(_xml, testXml);
         }
 
         [Test]
-        public void TRuleAllCondition_BuildXmlElement()
+        public void TRuleSMTPAuthCondition_BuildXmlElement()
         {
-            var testClass = new TRuleAllCondition(GetXmlNode(_xml));
+            var testClass = new TRuleSMTPAuthCondition(GetXmlNode(_xml));
 
             Assert.AreEqual(TRuleConditionType.None, testClass.ConditionType);
         }
